### 2025-08-19 #8833 @editorialbot
[PRE REVIEW]: Qualpal: Qualitative Color Palettes for Everyone
**Submitting author:** @jolars (<a href="http://orcid.org/0000-0002-4029-5945">Johan Larsson</a>)
**Repository:** https://github.com/jolars/qualpal
**Branch with paper.md** (empty if default branch):
**Version:** v3.2.0
