### 2025-08-19 #8833 @danielskatz
👋 @jolars - thanks for your submission.  As it's not completely clear to me that this is [research software](https://joss.readthedocs.io/en/latest/submitting.html#what-we-mean-by-research-software) as defined by JOSS, the editors will now discuss this. You should hear back in a week or two.
