### 2025-08-22 #8833 @jolars
> [@jolars](https://github.com/jolars) - this has passed the JOSS scope review, so it is ready to start, but we don't currently an appropriate available editor, so I'm going to put it on our waiting list.

Great to hear! Thanks for the quick triage!
