### 2025-09-04 #8833 @danielskatz
👋  @hollowscene & @dmreagan, would one or both of you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html
