### 2025-09-05 #8833 @ranzhengcode
﻿
﻿

Thank you for the kind invitation to review this manuscript. I am honored by your request. However, due to my current professional commitments, I am unable to take on any additional review tasks at this time. I apologize for any inconvenience and appreciate your understanding.

在 2025年9月5日，16:43，Verena Praher ***@***.***> 写道：

﻿
[https://avatars.githubusercontent.com/u/2265475?s=20&v=4]expectopatronum left a comment (openjournals/joss-reviews#8833)<https://github.com/openjournals/joss-reviews/issues/8833#issuecomment-3257564158>

👋 @sumedhsakdeo<https://github.com/sumedhsakdeo> & @expectopatronum<https://github.com/expectopatronum>, would one or both of you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html

Sorry for the late response! Unfortunately, I am currently involved in a very length review, and I cannot review this submission.
