### 2025-09-05 #8833 @rrrlw
Happy to help, if below considerations are acceptable

I cannot initiate review for next 2 weeks; I am not sufficiently qualified to review the standalone C++ library (happy to review R package & underlying C++ code), would request the other reviewer take lead on the standalone C++ library

Would this work?
