+++ .github/workflows/codecov.yml
@@ -35,6 +35,6 @@ jobs:
         run: ctest --test-dir build --output-on-failure
 
       - name: Upload coverage reports to Codecov
-        uses: codecov/codecov-action@v5.4.3
+        uses: codecov/codecov-action@v5.5.1
         env:
           CODECOV_TOKEN: ${{ secrets.CODECOV_TOKEN }}
