### 2025-09-09 8ac85b2
docs: remove faulty example
