### 2025-09-09 6bcbf83
docs: remove unncessary example
