### 2025-09-09 f3af5a7
docs: make analyze example straightforward
