### 2025-09-09 008f4e1
docs: use `-n 2` in example
