### 2025-09-09 b61435e
docs: complete installation instructions
