### 2025-09-09 d385dfa
docs: improve wasm instructions
