### 2025-09-09 be26461
docs: `setBackgroundColor` -> `setBackground`
