### 2025-09-09 93bebdc
docs: `setBackgroundColor` -> `setBackground`
Change this in a few more places.
