### 2025-09-09 7b9dc2f
docs: change example to directly use `analyze` subcommand
