### 2025-09-09 1150d52
docs: fix input palette example
