### 2025-09-09 ee5166b
feat(cli): honor `--metric` option in `analyze` subcommand
