### 2025-09-09 8a22814
chore: remove commented out code
