### 2025-09-09 9f8713d
feat(web): add controls to show colors simulated for CVD
Add a new UI component that controls how colors are shown, enabling
users to simulate how someone with CVD would see them.

refactor(web): let analysis component respond to CVD
