### 2025-09-23 f5afb8d
fix: use `-n 2` in CLI example
Thanks @clauswilke. Ref:
https://github.com/openjournals/joss-reviews/issues/8936
