### 2025-09-23 3bfe0d4
fix: ensure CLI output ends with newline
Ref: https://github.com/openjournals/joss-reviews/issues/8936. Thanks
@clauswilke
