### 2025-09-23 4daa2ca
chore: remove unnecessary comment
