### 2025-09-23 aef68ef
chore: remove unneccesary comment
