### 2025-09-23 0e80265
build: provide uninstallation target
Ref: https://github.com/openjournals/joss-reviews/issues/8936, thanks
@clauswilke
