+++ CMakeLists.txt
@@ -337,4 +337,21 @@ if(BUILD_WASM)
     )
 endif()
 
+# Uninstall target
+if(NOT TARGET uninstall)
+    configure_file(
+        "${CMAKE_CURRENT_SOURCE_DIR}/cmake/cmake_uninstall.cmake.in"
+        "${CMAKE_CURRENT_BINARY_DIR}/cmake_uninstall.cmake"
+        IMMEDIATE
+        @ONLY
+    )
+
+    add_custom_target(
+        uninstall
+        COMMAND
+            ${CMAKE_COMMAND} -P
+            ${CMAKE_CURRENT_BINARY_DIR}/cmake_uninstall.cmake
+    )
+endif()
+
 install(FILES LICENSE DESTINATION ${CMAKE_INSTALL_DOCDIR})
