### 2025-09-23 65ad76f
docs: update readme with prefix install and uninstallation
Ref: https://github.com/openjournals/joss-reviews/issues/8936, thanks
@clauswilke
