### 2025-09-24 f3c12c9
docs: correctly state ciede2000 as default color metric
