### 2025-09-24 69923fd
docs: show images of resulting palettes in readme and docs
Ref: https://github.com/openjournals/joss-reviews/issues/8936
