### 2025-09-24 6809d89
chore: add uninstall task
