### 2025-09-24 #8936 @jolars
> [@jolars](https://github.com/jolars) I think the paper is ready for publication in principle. Just one more comment: It would be nice to see some example color palettes featured somewhere. The paper doesn't show a single palette. The README on github doesn't show a single palette. Even the examples section in the documentation (https://jolars.github.io/qualpal/examples.html) doesn't show any palettes. My recommendation would be to add some in all these places.

Thanks, yes, sorry for this omission. I've updated the readme and docs (although the latter will only show with the next release). I'll update the paper soon too.
