### 2025-09-24 902f546
feat(cli): add `--extend` to support extending palettes
