### 2025-09-24 a37d139
docs: update readme and docs with additional examples
