### 2025-09-24 c252877
docs(paper): add an examples section to the paper
