+++ paper/paper.bib
@@ -158,6 +158,18 @@
   keywords      = {algorithm,color,farthest-points}
 }
 
+@online{okabe2008,
+  title         = {
+    Color Universal Design ({{CUD}}): {{How}} to Make Figures and Presentations That Are Friendly
+    to Colorblind People
+  },
+  author        = {Okabe, Masataka and Ito, Kei},
+  url           = {https://jfly.uni-koeln.de/color/},
+  urldate       = {2025-09-24},
+  type          = {J\textasteriskcenteredFLY},
+  date          = {2008-09-24}
+}
+
 @article{cui2002,
   title         = {Uniform Colour Spaces Based on the {{DIN99}} Colour-Difference Formula},
   author        = {Cui, G. and Luo, M. R. and Rigg, B. and Roesler, G. and Witt, K.},
