### 2025-09-24 0714e86
docs(paper): fix asterisk appearing in bibliography
