+++ paper/paper.bib
@@ -166,7 +166,7 @@
   author        = {Okabe, Masataka and Ito, Kei},
   url           = {https://jfly.uni-koeln.de/color/},
   urldate       = {2025-09-24},
-  type          = {J\textasteriskcenteredFLY},
+  type          = {JFLY},
   date          = {2008-09-24}
 }
 
