### 2025-09-24 #8936 @jolars
Okay, I have now updated the paper as well, adding a new section on examples, @clauswilke. Please let me know if you think that anything else is missing.

This upped the word count to 1314 (as per `pandoc paper/paper.md -t plain | wc -w`). I hope this is still fine. Please let me know otherwise.
