### 2025-09-24 8d5a938
chore(release): release 3.3.0 [skip ci]
## [3.3.0](https://github.com/jolars/qualpal/compare/v3.2.0...v3.3.0) (2025-09-24)

### Features

* **cli:** add `--extend` to support extending palettes ([902f546](https://github.com/jolars/qualpal/commit/902f546a220cad7d9c3a37e2df1afd70e0a1c747))
* **cli:** honor `--metric` option in `analyze` subcommand ([ee5166b](https://github.com/jolars/qualpal/commit/ee5166b401f6d1184545f0fdb5c2005288b864af))
* **web:** add controls to show colors simulated for CVD ([9f8713d](https://github.com/jolars/qualpal/commit/9f8713da07953d994247924c847747ec05d5bcf0))
* **w
[…]
