### 2025-09-24 1e44f74
docs: align code with output
