### 2025-09-24 a2e4054
docs: use title case for qualpal
