### 2025-09-24 5862c0d
docs: use standard text for package names in prose
