### 2025-09-25 85c7498
docs: add zenodo badge
