### 2025-10-14 f8c6f3d
Update paper.md
