### 2025-10-15 #8936 @jolars
Great to hear, thanks!

Tagged release: v3.3.0
DOI: 10.5281/zenodo.17194762

I've checked and updated the metadata on Zenodo.
