### 2025-10-17 #8936 @danielskatz
Congratulations to @jolars (Johan Larsson) on your JOSS publication!!

And thanks to @dmreagan & @clauswilke for reviewing!
JOSS is completely dependent on volunteers and we couldn't do this without you
